重大更新：SB 服修改模式，修改表结构

+++ src/main/java/com/now/nowbot/entity/BeatmapStarRatingCache.kt
@@ -22,7 +22,7 @@ class BeatmapStarRatingCache(
 
     var star: Double
 ) {
-    constructor(bid:Long, mode: OsuMode, mods: Int, star:Double): this(bid, mode.modeValue, mods, star)
+    constructor(bid:Long, mode: OsuMode, mods: Int, star:Double): this(bid, mode.modeValue.toShort(), mods, star)
     data class BeatmapStarRatingKey(
         var id: Long,
         var mode: Short,
